Reference

GenLib.branching — Function

branching(pedigree::OrderedDict{Int64, Individual}; pro::Vector{Int64} = pro(pedigree), ancestors::Vector{Int64} = founder(pedigree))

Return a pedigree that filters individuals who are in the paths between select probands and ancestors.

GenLib.children — Function

children(pedigree::OrderedDict{Int64, Individual}, ID::Int64)

Return the children of an individual.

GenLib.findDistance — Function

findDistance(pedigree::OrderedDict{Int64, Individual}, descendantID::Int64, ancestorID::Int64)

Return the minimum distance between an individual and their ancestor.

findDistance(genealogy::OrderedDict{Int64, Individual}, ID₁::Int64, ID₂::Int64, ancestorID::Int64)

Return the distance between two individuals and their ancestor.

GenLib.findFounders — Function

findFounders(pedigree::OrderedDict{Int64}, IDs::Vector{Int64})

Return a vector of founders from whom the IDs descend.

GenLib.findMRCA — Function

findMRCA(pedigree::OrderedDict{Int64, Individual}, IDs::Vector{Int64})

Return a vector of individuals' most recent common ancestors (MRCAs).

GenLib.founder — Function

founder(pedigree::OrderedDict{Int64, Individual})

Return a vector of founder IDs in alphabetical order.

GenLib.gc — Function

gc(pedigree::OrderedDict{Int64, Individual}; pro::Vector{Int64} = pro(pedigree), ancestors::Vector{Int64} = founder(pedigree))

Return a matrix of the genetic contribution of each ancestor (columns) to each proband (rows).

GenLib.genea140 — Constant

Genealogical information for 140 individuals from the Quebec Reference Sample.

According to the R GENLIB documentation, genea140 corresponds to "a genealogical corpus made of 41523 individuals from the province of Quebec, Canada. A total of 140 individuals have been sampled in seven sub-populations, listed in pop140, and their genealogies were reconstructed as far back as possible using the BALSAC population register and the Early Quebec Population Register."

GenLib.geneaJi — Constant

A highly inbred pedigree.

According to the R GENLIB documentation, geneaJi corresponds to "a modified version of a pedigree of two Jicaque Indians studied by Chapman & Jacquard (1971)."

GenLib.genealogy — Function

genealogy(dataframe::DataFrame)

Return an ordered pedigree of individuals from a DataFrame.

genealogy(filename::String)

Return an ordered pedigree of individuals from a CSV file.

GenLib.genout — Function

genout(pedigree::OrderedDict{Int64, Individual}, sorted::Bool = false)

Return a pedigree as a DataFrame.

If sorted is false (the default), then the individuals will appear in the same order as in the pedigree.

If sorted is true, then the individuals will appear in alphabetical ID order.

GenLib.occ — Function

occ(pedigree::OrderedDict{Int64, Individual}; pro::Vector{Int64} = pro(genealogy), ancestors::Vector{Int64} = founder(genealogy), typeOcc::String = "IND")

Return a matrix of ancestors' occurrences.

If typeOcc is :ind (default), then the matrix corresponds to the occurrence per individual. If typeOcc is :total, then the matrix corresponds to the total occurrence.

GenLib.phi — Function

phi(individualᵢ::Individual, individualⱼ::Individual, Ψ::Union{Nothing, Matrix{Float64}} = nothing)

Return the kinship coefficient between two individuals.

A matrix of the founders' kinships may optionally be provided.

phi(pedigree::OrderedDict{Int64, Individual}, rowIDs::Vector{Int64}, columnIDs::Vector{Int64})

Return a rectangle matrix of kinship coefficients between row IDs and column IDs. The kinship of someone with themself is replaced with their inbreeding.

phi(pedigree::OrderedDict{Int64, Individual}, Ψ::Matrix{Float64})

Return a square matrix of pairwise kinship coefficients between all probands given the founders' kinships.

An implementation of the recursive-cut algorithm presented in Kirkpatrick et al., 2019.

phi(pedigree::OrderedDict{Int64, Individual}, probandIDs::Vector{Int64} = pro(pedigree); verbose::Bool = false)

Return the square matrix of the pairwise kinship coefficients of a set of probands.

If no probands are given, return the square matrix for all probands in the pedigree.

An implementation of the recursive-cut algorithm presented in Kirkpatrick et al., 2019.

GenLib.pop140 — Constant

Population of origin of the 140 Quebec samples.

According to the R GENLIB documentation, pop140 corresponds to "140 individuals from the genealogical corpus from Quebec (…) sampled from 7 different opulations from 5 regions: Quebec City, Montreal, Saguenay, North Shore, Gaspesia. In Gaspesia we find 3 different populations: French-Canadians, Acadians and Loyalists."

GenLib.pro — Function

pro(pedigree::OrderedDict{Int64, Individual})

Return a vector of proband IDs in alphabetical order.

GenLib.rec — Function

rec(pedigree::OrderedDict{Int64, Individual}, probandIDs::Vector{Int64} = pro(genealogy), ancestorIDs::Vector{Int64} = founder(genealogy))

Return the number of descendants of each ancestor.