Reference

GenLib.genea140 — Constant

Genealogical information for 140 individuals from the Quebec Reference Sample.

According to the R GENLIB documentation, genea140 corresponds to "a genealogical corpus made of 41523 individuals from the province of Quebec, Canada. A total of 140 individuals have been sampled in seven sub-populations, listed in pop140, and their genealogies were reconstructed as far back as possible using the BALSAC population register and the Early Quebec Population Register."

Loading the Dataset

import GenLib as gen
genea140 = gen.genea140
ped = gen.genealogy(genea140)

GenLib.geneaJi — Constant

A highly inbred pedigree.

According to the R GENLIB documentation, geneaJi corresponds to "a modified version of a pedigree of two Jicaque Indians studied by Chapman & Jacquard (1971)."

Loading the Dataset

import GenLib as gen
geneaJi = gen.geneaJi
ped = gen.genealogy(geneaJi)

GenLib.pop140 — Constant

Population of origin of the 140 Quebec samples.

According to the R GENLIB documentation, pop140 corresponds to "140 individuals from the genealogical corpus from Quebec (…) sampled from 7 different opulations from 5 regions: Quebec City, Montreal, Saguenay, North Shore, Gaspesia. In Gaspesia we find 3 different populations: French-Canadians, Acadians and Loyalists."

Loading the Dataset

import GenLib as gen
pop140 = gen.pop140
pop = gen.population(pop140)

GenLib.GenMatrix — Type

struct GenMatrix
    individuals::Vector{Int64}
    ancestors::Vector{Int64}
    meioses::Matrix{Int64}
end

A matrix that goes with individuals as rows and ancestors as columns.

GenLib.Individual — Type

struct Individual <: AbstractIndividual
    ID::Int64
    father::Union{Nothing, Individual}
    mother::Union{Nothing, Individual}
    children::Vector{Individual}
    sex::Int64
    rank::Int64
end

The unit structure of a Pedigree.

GenLib.Pedigree — Type

const Pedigree{T} = OrderedDict{Int64, T} where T <: AbstractIndividual

A particular case of an OrderedDict containing individuals accessed by ID.

GenLib.branching — Function

branching(pedigree::Pedigree; pro::Vector{Int64} = pro(pedigree), ancestors::Vector{Int64} = founder(pedigree))

Return a pedigree that filters individuals who are in the paths between select probands and ancestors.

GenLib.children — Function

children(pedigree::Pedigree, ID::Int64)

Return the children of an individual.

GenLib.findDistance — Function

findDistance(pedigree::Pedigree, IDs::Vector{Int64}, ancestorID::Int64)

Return the distance between two individuals and their ancestor.

GenLib.findFounders — Function

findFounders(pedigree::Pedigree, IDs::Vector{Int64})

Return a vector of founders from whom the IDs descend.

GenLib.findMRCA — Function

findMRCA(pedigree::Pedigree, IDs::Vector{Int64})

Return a GenMatrix of meioses between individuals and their most recent common ancestors (MRCAs).

Example

import GenLib as gen
genea140 = gen.genea140
ped = gen.genealogy(genea140)
pro = gen.pro(ped)
pro1 = pro[1]
pro2 = pro[2]
genMatrix = gen.findMRCA(ped, [pro1, pro2])

GenLib.founder — Function

founder(pedigree::Pedigree)

Return a vector of founder IDs in alphabetical order.

Example

import GenLib as gen
genea140 = gen.genea140
ped = gen.genealogy(genea140)
founders = gen.founder(ped)

GenLib.gc — Function

gc(pedigree::Pedigree; pro::Vector{Int64} = pro(pedigree), ancestors::Vector{Int64} = founder(pedigree))

Return a matrix of the genetic contribution of each ancestor (columns) to each proband (rows).

Example

import GenLib as gen
genea140 = gen.genea140
ped = gen.genealogy(genea140)
contributions = gen.gc(ped)

GenLib.genealogy — Function

genealogy(dataframe::DataFrame; sort::Bool = true)

Return an ordered pedigree of individuals from a DataFrame.

Example

import GenLib as gen
using DataFrames
inds = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
fathers = [0, 0, 0, 1, 1, 0, 3, 3, 6, 6]
mothers = [0, 0, 0, 2, 2, 0, 4, 4, 5, 5]
sexes = [1, 2, 1, 2, 2, 1, 2, 1, 1, 2]
df = DataFrame([inds, fathers, mothers, sexes], [:ind, :father, :mother, :sex])
ped = gen.genealogy(df)

genealogy(filename::String)

Return an ordered pedigree of individuals from a CSV file.

Example

import GenLib as gen
genea140 = gen.genea140
ped = gen.genealogy(genea140)

GenLib.genout — Function

genout(pedigree::Pedigree, sorted::Bool = false)

Return a pedigree as a DataFrame.

If sorted is false (the default), then the individuals will appear in the same order as in the pedigree.

If sorted is true, then the individuals will appear in alphabetical ID order.

import GenLib as gen
geneaJi = gen.geneaJi
ped = gen.genealogy(geneaJi)
gen.genout(ped)

GenLib.occ — Function

occ(pedigree::Pedigree; pro::Vector{Int64} = pro(genealogy), ancestors::Vector{Int64} = founder(genealogy), typeOcc::String = "IND")

Return a matrix of ancestors' occurrences.

If typeOcc is "IND" (default), then the matrix corresponds to the occurrence per individual. If typeOcc is "TOTAL", then the matrix corresponds to the total occurrence.

Example

import GenLib as gen
geneaJi = gen.geneaJi
ped = gen.genealogy(geneaJi)
occ = gen.occ(ped, typeOcc = "TOTAL")

GenLib.phi — Function

phi(individualᵢ::T, individualⱼ::T) where T <: AbstractIndividual

Return the kinship coefficient between two individuals.

Adapted from Karigl, 1981.

Example

import GenLib as gen
genea140 = gen.genea140
ped = gen.genealogy(genea140)
pro1 = ped[10033]
pro2 = ped[113470]
gen.phi(pro1, pro2)

phi(individualᵢ::PossibleFounder, individualⱼ::PossibleFounder, Ψ::Matrix{Float64})

Return the kinship coefficient between two individuals given a matrix of the founders' kinships.

Adapted from Karigl, 1981, and Kirkpatrick et al., 2019.

phi(pedigree::Pedigree, Ψ::Matrix{Float64})

Return a square matrix of pairwise kinship coefficients between all probands given the founders' kinships.

An implementation of the recursive-cut algorithm presented in Kirkpatrick et al., 2019.

phi(pedigree::Pedigree, probandIDs::Vector{Int64} = pro(pedigree); MT::Bool = false)

Return a square matrix of pairwise kinship coefficients between probands.

If no probands are given, return the square matrix for all probands in the pedigree.

An implementation of the recursive-cut algorithm presented in Kirkpatrick et al., 2019.

Example

import GenLib as gen
geneaJi = gen.geneaJi
ped = gen.genealogy(geneaJi)
gen.phi(ped)

phi(pedigree::Pedigree, rowIDs::Vector{Int64}, columnIDs::Vector{Int64})

Return a rectangle matrix of kinship coefficients, as defined by a list or row IDs and column IDs.

GenLib.pro — Function

pro(pedigree::Pedigree)

Return a vector of proband IDs in alphabetical order.

Example

import GenLib as gen
genea140 = gen.genea140
ped = gen.genealogy(genea140)
probands = gen.pro(ped)

GenLib.rec — Function

rec(pedigree::Pedigree, probandIDs::Vector{Int64} = pro(genealogy), ancestorIDs::Vector{Int64} = founder(genealogy))

Return the number of descendants of each ancestor.